updated custom table node tree
